fix 프로젝트 구조 변경 및 데이터 불러오기 async함수로 변경
