+++ src/App.js
@@ -1,25 +1,19 @@
-import logo from './logo.svg';
-import './App.css';
-
+import "./style/App.css";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+import Item from "./pages/Item";
+import AddItem from "./pages/AddItem";
 function App() {
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
+    <BrowserRouter>
+      <div className="App">
+        <Routes>
+          <Route path="/items/:id" element={<Item />} />
+          <Route path="/additem" element={<AddItem />} />
+        </Routes>
+      </div>
+    </BrowserRouter>
   );
 }
 
 export default App;
+
